Update fix regulation bug
This changes made to this commit is to fix a bug when there are not
tokens or words, we try to allocate memory for Zero words and then try
to write or free it.

The fix is, if there are no words we can safely return as there is
nothing to allocate.